with regard to #1, we can now accurately work with a limited dataset.
With this commit the only thing we need to know in order to be able to
determine flight times:

- Latitude
- Longitude
- Timestamp

This will allow us to combine FLARM feeds with ADS-B feeds in order to
get a more accurate picture of what's going on.

But please note we need something to figure out timing issues when mid
flight. Those are a PITA.