Fixed some issues responsible for not processing data at all
In case a limited subset of data would be used data would not be
processed at all, and an unhandled exception would be thrown.

Currently there's still an issue related to processing of data in
limited amounts. If the update speed at which we get our data is higher
than the speed with which new GPS fixes are being retrieved, we see that
as an aircraft that's standing still.

We could fix that by:
- Calculating G load. If it's higher than 9G we could either ignore it
or mark the crash site.
- Just let it be, and only mark an arrival if there are multiple points
in a row (like 5 or something) that say the aircraft has a groundspeed
of 0.